Web: Work harder to turn inaccessible links into something presentable
We were already parsing the URL of an inaccessible link in order to
present something besides just "link" when one of these gains focus
(e.g. via Tab). But we were checking to see if the basename we
computed for the URL consisted only of alphabetic characters. This
was done in order to filter out long gibberishy URLs like we see
associated with ads. Being this strict, however, meant that a basename
of "foo" would be presented, whereas a basename of "foo-bar" would not.
Therefore, split the basename on "-" and "_" into a series of tokens.
If each of those tokens consists solely of alphabetic characters,
present it.