Web: Handle caret-moved event preceding focus event when Tab pressed in entry
Chrome/Chromium appears to be emitting a caret-moved event before the
focus event the first time a new input/entry gains focus as a result
of tabbing into it. When this occurs, we update the locusOfFocus but
wind up presenting nothing because tabbing into editable text is not
a case handled by the default script.

In this particular case, we know that we will be getting a focus
event from the browser (and if we don't, we need that fixed in the
browser). So it's safe to ignore this initial caret-moved event and
wait for the focus event to do the presentation.