Chromium: Work around bogus/misleading selected state in browser menus
Once one arrows past zoom-related or edit-related items in the three-dot
menu, zoom and edit each claim to be selected. Because AtkSelection does
not report the selected children for this menu, we resort to examining
all children to find those with state-selected and wind up with two:
the real selected child and zoom or edit. All this needs to be fixed in
Chromium, but it appears that catching this condition and sanity-checking
items by also examining the focused state is a safe workaround in the
meantime.