Chromium: Fix bogus presentation of posinset and setsize
Chromium has accessible menu items which are not focusable and therefore
do not have a posinset and setsize calculated. But they may claim to be
the selected item when an accessible child is selected (e.g. "zoom" when
"+" or "-" gains focus. Normally we calculate posinset and setsize when
the application hasn't provided it. We don't want to do that in the case
of menu items like "zoom" because our result will not jibe with the
values of its siblings. Thus if a sibling has a value, assume that the
missing attributes are missing on purpose.