Chromium: Fix chattiness arrowing up/down in omnibox popup
We were resetting the locusOfFocus in response to caret-moved events
received with each selection change. These events should be ignored
when arrowing up/down in the popup for a single-line autocomplete.

This change is being made in the web script rather than the Chromium-
specific support because this condition could happen in Firefox too.