Web: Prevent getting stuck when up arrowing with multiple newline chars
The presence of multiple newline characters in content can lead to us
getting the same line via the accessible text interface. If we don't
sanity check and handle this condition, the user can get stuck in
content. We were handling this condition for moving forward by line,
but not backward by line. Apply the same strategy, namely if we get
the same line for a previous offset decrement the offset again and
try once more.