Be less strict about which input events we treat as Tab navigation
At least some users have Tab input events for which there is a modifier
we don't recognize. As a result, we can fail to recognize Tab/Shift+Tab
navigation which can impact what we do or do not present. Therefore,
just filter out things we don't wish to treat as Tab navigation, such
as Alt, Ctrl, and the Orca modifier.