Web: Improve logic identifying text brokenness due to it being off-screen
When authors hide text just for screen-reader users to find, it is
often exposed to us via the accessible text interface as if there were
only one character per line. We have logic to check for this case, but
it wasn't catching things like:

* our getting a valid line when the first character is an embedded
  object, but subsequent characters are not
* our getting a character along with a space (two characters are
  not a single character)

Therefore add handling to ignore results when the character at offset
is embedded object, and strip whitespace before examining the supposed
line at offset.