Web: Handle yet more off-screen text-interface brokenness
It turns out that we sometimes get non-broken text when the off-screen
text at offset happens to be a space, but for all other offsets we
get just a single-character per line. Therefore also ignore spaces
when doing the sanity check. Ditto for the replacement character. We
can see the latter in some implementations when we use getText() for
a length of one at the offset of a multibyte character. In theory,
we should instead call getTextAtOffset() asking for a character. But
due to other bugs in (at least) Gecko, that is even less reliable.