Web: Prevent getting stuck/looping in offscreen single-char-lines text
When we search for the first valid caret position at/after a specified
offset, we typically descend the element if the character at offset is
an embedded object character. In the case of offscreen text which is
exposed to us as one character per line, this can lead to getting stuck
or looping because we always return a range of 0:length for the broken
element.