Handle massively-delayed selection-changed events for apps like Caja
Normally when a user copies and pastes a file/folder in a file manager,
we get a selection-changed event from the container of the just-pasted
item and want to update our location. But when copying and pasting a
large number of items in Caja, we are placed in a new window for the
progress and eventually get the aforementioned selection-changed event.
Therefore, be sure the container is in the active window before updating
our location.