Web: Don't use character count to determine text-block-elementness
If an object is in the process of being destroyed, getting its
character count (without exception handling) can cause Orca to spit
up. While we could of course add that exception handling, the lack
of text content shouldn't cause us to deem an element not texty; it
is simply an empty text element. Therefore remove that character-
count check entirely.