Event manager: Allow scripts to handle locusOfFocus destruction
When the locusOfFocus is being destroyed, the event.any_data from the
children-removed event sometimes contains an already-dead accessible
object. We were ignoring these events on the grounds that there wasn't
much to do other than wait for an event (e.g. focused) for the new
location. But scripts which cache objects (e.g. web) might be able
to do some cleanup, so allow that to happen by not ignoring these
events.