Event Manager: Try to detect and avoid selection-changed event flood
Apps like Caja seem to like to flood us with active-descendant-changed
and selection-changed events. We are already ignoring the former based
on the supposed new active descendant being defunct. We were also
trying to catch this condition and ignore the selection-changed flood,
but were returning early when the active-descendant was already dead.
Its death short-circuited the remaining logic. Therefore, don't return
quite so early for dead descendants so that we can add the spammy
offender to the list of objects destroying children.