Web: Further limit when we generate the line for editable with descendants
The previous commit ensured we spoke just the newly-focused editable
object when Tab was pressed. But the object could be autofocused by
the app, or given focus by mouse click. Therefore only generate the
full line for newly-focused editable objects when Orca's caret nav
or structural nav was used to move into them.