Event Manager: Allow scripts to handle child removal if focus is dead
Sometimes we get a single event for a child removal, but the child
being removed is potentially an ancestor of the locusOfFocus. Or
there may be a case where the equality check between the removed child
and the locusOfFocus fails. Therefore if the locusOfFocus is dead,
allow scripts to handle this event so they can do any possibly clean
up / recovery.