Prevent infinite recursion generating statusbar content
It appears that sometimes we're getting the status bar itself as part
of the status bar's items (or perhaps the status bar has a descendant
status bar??). This can lead to infinite recursion. Therefore, filter
out any status bars returned when we get the status bar's items. Also
add a redundant check to the speech generator to ensure we avoid this
problem.