Web: When caret context search fails in document return the document itself
When trying to locate the caret location in a newly-loaded page, we rely
on the accessible text interface being implemented on the document. If
for some reason (e.g. a Firefox regression) that isn't the case, it's
better to return the document itself. Other logic should then make it
possible to determine what to present.