Prevent double-presentation of description for alerts
In (at least) Firefox's native alerts, the alert text is exposed as
the accessible description. Knowing this, that's our last resort for
the displayed text. But when we rely upon that and the user also has
descriptions spoken (which they are by default), we double-present
the text.

Also ensure this last resort kicks in even if the user has disabled
spoken descriptions.