LibreOffice: Work around missing state-changed:focused event
After (at least) leaving the Save As dialog, paragraphs claim to be
focusable but no longer claim to be focused. As a result, we ignore
text-related events from those paragraphs having concluded the user
is not interacting with them. Clearing the cache for these paragraphs
seems to get them to report the focused state when they have it. So
it seems LibreOffice is failing to emit the state-changed event needed
to get AT-SPI2 to update its cache.