Web: Stop initially on giant ARIA containers in browse mode
Orca used to not descend certain container roles in browse mode, such
as menus, toolbars, trees, etc. The reason being we trusted that the
author would provide focus management for these items and the user
could access them in focus mode. Unfortunately, we found instances in
which the author ARIAed up static content with these roles and failed
to provide any focus management so Orca now descends these items in
browse mode.

Unfortunately, this change seems to have a side effect of chattiness
in certain properly-authored content, such as an SVG with ARIA tree-
related roles, proper names, and optional author-provided navigation.

Because we cannot know programmatically if we have good authoring or
bad authoring, a reasonable compromise seems to be to initially stop
at these large containers and present them to the user in browse mode.
This gives the user the opportunity to manually enter focus mode if
desired and also eliminates some verbal spew if they remain in browse
mode.