Web: Eliminate some chattiness entering elements with a caption
If an element has a caption there is an excellent chance that its
accessible name will be derived from it. If that caption is visible/
navigable, we'll likely end up presenting it again during navigation.
In addition, captions (e.g. on figures and tables) can get rather long.
The least chatty thing to do in this case seems to be to not present
visible captions as the name but rather wait until the user navigates
to it.