Web: Prevent trees and tree items from being presented on the same line
Trees can be read via browse mode if the user opts to do so. However,
depending on the associated markup, the items might be rendered on
the same line. Even in browse mode, we don't want to treat trees and
tree items as if they are on the same line beacuse it can be confusing
and chatty.