Web: Make ARIA treeitems descendable in browse mode
Now that trees can be navigated in browse mode (mainly to work around
bad authoring), we need to conditionally descend treeitems in case we
have an instance of nested treeitems. This change in and of itself is
simple. However, if instead of nested treeitems, the descendants are
the non-interactive text and/or image elements used to display the
items content, we don't want to descend into them.

It's not performant to do a tree dive just to check which kind of
tree we have. Therefore, assume the treeitem is descendable, but
treat its non-interactive descendants as non-valid caret contexts.