Web: Improve location recovery using interactive trees in browse mode
If the user expands/collapses an ARIA tree item, there will likely be
children-changed events which will cause us to dump our cache. When
the user is in focus mode, subsequent focus/selection events emitted
by the web app will make it possible for us to update the caret context
so we know where we are with respect to the page. In browse mode, we
likely won't get those events. Therefore, update the caret context when
we get an expanded-changed event for the locusOfFocus and do not have
a cached caret context. This should keep us from jumping to the wrong
location (i.e. outside of the tree) in response to the next arrow press.