Web: Don't search for caret context in a hidden element
If an element is hidden, it and all of its descendants cannot have
the caret context. So there's no point in searching there.

In addition, this *may* fix an instance of infinite recursion in
which we try to recover from the focused accessible element being
removed as a side effect of its parent becoming hidden.