Evolution: Work around missing events from Evolution Add Accounts
When the "Add" button is pressed in the Evolution Accounts dialog,
a popup menu appears and claims focus. But when we check if the
top-level window containing that menu has state active, it does not.
This causes us to reject the event, along with subsequent navigation
in the menu.

It seems safe enough to handle this specific case for Evolution by
trusting that menus claiming focus are in the active window even if
that window claims to not be active.