Web: Fix braille presentation of grid cell in browse mode
The web braille generator was defering to the default braille generator
for presentation of grid cell content when in browse mode. We stopped
doing that for speech so that additional web-specific logic could be
used to ensure we presented something reasonable. We should have done
the same for braille. This commit makes that change so that we're
speaking and brailling the same objects for grid cells in browse mode.