Be more flexible with the accessibility tree in _generateUnselectedCell
Assuming that the parent of a cell is a table is a holdover from the
pre-ARIA days. Rows may or may not be present in the accessibility
tree and there may also be other containers, such as a row group.
Therefore, ascend the accessibility tree looking for a table ancestor.