Web: Ensure we dump content cache for editable text deletions
We were not dumping our content cache in editable-text elements. Then,
in order to ensure we didn't present stale information, we didn't use
the cached content when speaking the word or line at offset if the
element was editable. This can increase lagginess in rich text editors
(e.g. Google Docs) in response to caret-moved events when braille is
enabled because we build the line twice: once for braille and once
for speech of an editable element. We can eliminate some lagginess by
ensuring we dump the content cache for text deletions, and then using
the updated cache when we speak the new position.