Web: Make onCaretMoved more efficent
We need to do a number of checks to determine if and how every caret-moved
event should be presented. Many of these checks were causing us to seek
the document frame, and often doing so from the top down which can be
less performant than searching the ancestry. Therefore, get the top-level
document for the event source and include it as an argument where possible.