Web: Ensure we present caret-moved events after Alt+Tabbing into web app
One of the checks we do when determining whether or not to present a
caret-moved event is see if it came from the current document by
getting the ancestor document of the locusOfFocus and the ancestor
document of the event source. If the user Alt+Tabbed into an application
and no focus event was emitted, the locusOfFocus will be the window
itself and not have any ancestor document. This will cause presentation
of caret-moved events to fail.

To handle this condition, we should use the active document as the
locusOfFocus document when the locusOfFocus is a window. This, however,
may not be enough to fix the bug: The active document is retrieved via
the embeds relationship on the window or, failing that being implemented,
from the top down. As a result, we can still get a document mismatch
when the event source is inside a nested document (which can happen in
web apps such as VSCode). Therefore, when the locusOfFocus is the active
window, we also need to retrieve the topmost document of the event source
to ensure we're comparing the right objects.