Web: Fix regression presenting plain-text messages in Thunderbird
getTopLevelDocumentForObject() might be passed an object which is the
top level document, such as when we get a caret-moved event in a plain-
text document in certain versions of Thunderbird. Assuming that caret-
moved events will be emitted from a descendant of the document can
result in our concluding the event is not from document content.