Treat state-changed:active true on frames as reason to change script
The event manager was treating window:activate as a reason to switch
the active script. But it was not doing so for state-changed:active
on a frame. As a result, if an application emits the latter before
the former -- or fails to emit the former at all -- script-specific
logic might not kick in for the activation.

This event ordering was causing us to switch to browse mode incorrectly
in VSCode. Treating state-changed:active true on frames as a reason to
change the active script solves this problem.