Event Manager: Add initial handling for floods and deluges
Some applications can be incredibly spammy and flood us with events
that grind Orca to a halt. This commit adds initial flood and deluge
detection handling. It will undoubtedly need to be tweaked.

* Flood: We have more than 100 events in the event queue
* Deluge: We have more than 250 events in the event queue
* During a flood, there are some events which we will dequeue, but
  then not send to the script for processing. Rationale: At the time
  we enqueued them, things we're still good; since then things are
  bad and we should move on to more recent events.
* During a deluge, there are some events which we won't even bother
  enqueuing because things are just awful, and what's the point?

At the present time, the only events that are treated as ignorable are
text-changed:insert and text-changed:delete for objects that are not the
locusOfFocus. We always care about events from the locusOfFocus, but
text-changed spam from other objects (which can come from DOM mutations)
take too long to process.