Event Manager: Clear queue of spam during flood if given a priority event
This commit makes it possible to specify events that, if received during
an event flood, should trigger pruning of spam events.

At the present time, the only event which triggers pruning is a focus
claim. We might wish to expand that to include other events after we
see that there are no unexpected side effects from this change.