Event Manager: Don't switch scripts for activate on already-active window
If we receive a window:activate and/or state-changed:active event on a
window after we've received a focus event (or figured out focus on our
own), we don't want to switch scripts. While in most cases doing so has
no impact, there are times when the script for the window is not the
same as the script for the focused object. This is true for apps which
embed WebKitGtk and terminal applications.

Therefore, only switch applications when we get window activation for
a window which is not already the active window. This solves a recent
regression introduced by d261ce21e and should also solve a similar
problem which users report has been around a while.