Event Manager: Add state-changed:selected to spam clearers
This is similar to what we do with state-changed:focused. If the object
is a container which supports selection, we might not get a focus change,
but should get a selection change. This should also handle the case where
gnome-shell occasionally spams us with state-changed:showing events,
because it emits selection change events.