Web: Improve recovery from destruction of current object
When we received an event signaling the removal of the locusOfFocus,
we attempted recovery. We were not attempting this recovery if the
removed child was an ancestor of the locusOfFocus. This commit does
so and should make it more likely that we do not lose our place if
a web app destroys the active element.