Web: Fix bug in toolbar browse-mode navigation
If we're in a toolbar which is descendable (e.g. because we're in browse
mode), the line contents should be for the descendants and not include
the toolbar itself.

Including it as part of the contents can lead to our getting stuck when
we try to navigate past it due to a to-be-debugged issue. However, since
the toolbar itself shouldn't be on the line anyway, remove it.