LibreOffice: Work around missing focus event for Calc spreadsheet
After saving a document and returning to the Calc spreadsheet, there
is no object:state-changed:focused event for the table. As a result,
AT-SPI2 does not update its cache and we're given active-descendant-
changed events on an object which claims to be unfocused. Thus we
ignore them.

Therefore, check in the app script for this condition and clear the
cache on the table before handing the event off to the default script.