LibreOffice: Don't update locusOfFocus to dead object
In LibreOffice, when we get a text-selection-changed event from an
object which claims to be focused but is not the locusOfFocus, we
update the locusOfFocus before passing the event onto the default
script. We do this because LibreOffice isn't super reliable about
emitting focus events, and the default script will ignore the event
otherwise.

Unfortunately, Calc is emitting text-selection-changed on a destroyed
object which claims to be focused. Therefore, check if this object is
dead and return early if it is.