Web: Fix issues related to ARIA tooltips
If an ARIA tooltip claims to be focusable, we treat it as something
presentable even if it has no size and no content. This leads to
chattiness, and occasionally even setting the caret in, a useless
object. The latter can trigger focus events, along with object
creation and destruction that leads to looping within the content.
Solve both problems by not allowing an empty tooltip to have the
caret context.

Also fix a traceback checking for the name of a recently-destroyed
ARIA tooltip.