Web: Identify and work around mismatched text, hypertext implementations
The value of hyperlink.startIndex should match the offset in the text
implementation's character offset for the associated child object.
When this fails, due to a bug in the browser implementation, we can
get stuck/loop. Therefore, check for this mismatch, spew out plenty
of debugging details, and try to recover by using the hyperlink's
reported range instead of a character offset when we've gotten stuck
on a line.