Web: Handle caret-moved events preceding focus in spin buttons
Chrome (and possibly other browsers) emits caret-moved events prior to
focus events with inputs. This can cause us to quietly set the focus
and subsequently ignore the focus event. We were already handling that
ordering for inputs exposed as entries. But inputs with a type of
number get exposed as spin buttons. This commit fixes the issue for
numeric inputs.