Don't fall back accessible name for displayed text in spin button
When getting the displayed text of something which claims to implement
the accessible text interface but then doesn't, we fall back on the
name. That doesn't make sense for a spin button in general. In addition,
there seems to be a bug in Chrome/Chromium in which spin buttons which
display text fail to expose that text via the text interface. Falling
back on the name prevents other checks (e.g. valuetext property) from
happening. So this change in Orca will work around that issue in
Chrome/Chromium.