Web: Clear cached objects when not dumping full cache for child change
In order to be performant, we only dump our full cache in response to
children-changed events when that event is fired by the locusOfFocus
or one of its ancestors, or the locusOfFocus is defunct. However,
certain cached object properties may need to be updated in response
to children changing. For instance, if a nearby and formerly empty
div adds children, it is no longer "useless". If we don't clear such
properties, we might skip over content.