Web: Check for sizeless headings when building line contents
Normally we want to present visual line contents in layout mode.
But we include unrendered, sizeless objects on the same line
because they are presumably there for accessibility purposes and we
don't want their inclusion to break up the rest of the line. There is
at least one site including an unrendered, sizeless heading right
before the content. We should treat that heading, and those like it,
as the full line.