Improve presentation of all text becoming unselected during caret navigation
If text was selected and then a caret navigation command caused all that
text to become unselected, we were presenting the text which became
unselected and not speaking the new location in certain applications.
The former is chatty, and the latter not helpful. Therefore, when all of
the text becomes unselected, speak a short "selection removed" message.
Then present the new location.