Web: Attempt to work around object destruction during structural nav
Browsers sometimes destroy objects as a result of our grabbing focus
and/or setting the caret in those elements. When this happens, Orca
fails to present the new object correctly because it either no longer
exists at all, or lacks information it once had (like its name) due to
the destruction process having been started.

This commit attempts to work around that problem during structural
navigation by detecting the condition and then looking for a
replacement object at the same accessible path and with the same role.

Note that this work around sometimes fails due to timing and/or browser
bugs. More specifically, if AT-SPI2 has not received a children-changed
event from the parent of the element in question at the time we look for
that replacement, we'll still wind up with the defunct object we had
before. That said, this workaround seems to make things suck a bit less
than doing nothing at all to recover.