Only get the voice during speech generation if there's something to say
The speech generator code was frequently getting the voice prior to
generation and then applying it if there was something to say. Up to
now that was fine because the getting the voice was merely retrieving
it from settings. This will change when on-the-fly language switching
is fully implemented and enabled. Therefore for performance purposes,
we should only get the voice if we definitely have something to say.