Include the object when scripts retrieve the voice directly
Ideally all speech generation would be done by the speech generators.
Unfortunately, there is some legacy code in the scripts which bypasses
this logic. That code needs to include the object when retrieving the
voice from the speech generator so that the to-be-implemented on-the-fly
language switching works with those methods.