Web: Announce locusOfFocus update during object destruction
Some web apps destroy the focused object and replace it. We normally
handle this silently in order to not be chatty. Sometimes, however,
not saying anything can result in not presenting an object which is
about to claim focus. Our choice is to not try to recover and hope
we get a focus event from the browser (which might not occur), or
risk being chatty and presenting the object if it's focused. This
commit does the latter.