Web: More aggressively update focus for focused entries and spinners
Typically we get focus events prior to caret moved events for focused
entries and spinners. Except for when we don't. When we silently update
our location in cases where we don't, we are in danger of failing to
present a newly-focused entry. We're seeing this in VSCode. Therefore
move the handling which we were doing in response to pressing the Tab
key to the end of onCaretMoved(). If we get to the end and no other
case has addressed the situation, handle the location change by
presenting it.