Generate current line text even if there is an embedded object char
_generateCurrentLineText() has for many years returned early and
without generating anything if an embedded object character was
found. The idea appears to be that in those cases returning nothing
would trigger the logic needed to split such lines up (e.g. in the
case of web content). The web script, however, handles line generation
differently and does not rely upon _generateCurrentLineText(). So
returning early doesn't seem necessary. In addition, LibreOffice Writer
inserts embedded object characters for comments as well as images and
other objects. (Unfortunately its hypertext implementation is broken,
so reliably presenting the embedded objects like we do on the web is
not possible.)

This commit stops us from returning early and removes any embedded
object characters so that we can present the line's textual contents
as expected.