Event manager: Process children-changed from tables in async mode
Our flood detection and handling code assumes we are queuing events.
We typically don't queue children-changed so that we can try to recover
from object destruction (e.g. on the web). As a result, if an app or
toolkit decides to populate a giant table row by row, notifying us
as it goes, we'll grind to a halt.