Eliminate some chattiness from text insertions
We were treating text insertions which met a number of requirements
as possibly presentable if the locusOfFocus was dead. But this can
result in our presenting text we shouldn't, for instance when VSCode's
editor gets destroyed and recreated when the user shows the Problems
view.

We can eliminate some of this chattiness by checking for the focused
state: If we're not in it, and we don't (yet) know where we are, and
the object in which the text was inserted isn't focused, ignore it.