Web: Eliminate some chattiness from ARIA page tabs
The ARIA 'tab' role supports aria-expanded and nothing prohibits it
from being applied in a single-selection tablist. If an author sets
both aria-selected and aria-expanded on the newly-selected/active
tab, we'll double-present that tab. Therefore, store the expanded
state when selection changes so we can identify when it truly changed
independent of selection.