Improve logic to filter redundant names
The logic to filter out names which were redundant to labels handled
the following conditions:

* The name and label only differed with respect to whitespace
* The label started with the text of the name (e.g. because the
  author/developer wanted to append some helpful information)

It did not, however, handle the following:

* The name and label only differed with respect to punctuation
* The name itself had duplicate text (e.g. repeats the label twice)

This failure makes Orca quite chatty with Gtk 4 widgets where the
use_underline attribute is present because Gtk 4 exposes a name
consisting of the displayed text with the underline included
followed by that same text without the underline included.

This commit handles those two missing scenarios.

Fixes #237.