Filter out redundant described-by targets
Gtk 4 is sometimes exposing one or more targets both through the
labelled-by relation and the described-by relation. This causes us to
double-speak the information if the user has enabled the presentation
of descriptions.

This commit returns an empty list of described-by targets when they
are the same as the list of labelled-by targets in order to eliminate
that duplication.