Web: Ignore text selection changes in entries preceding focus/caret-moved
If we get a caret-moved event or focus-change event for a focused entry,
we update the locusOfFocus and then present the widget and text, including
any text selection. Right now we're automatically presenting text selection
changes in editable content which is not the locusOfFocus. This leads to
double-presentation because the caret-moved event and/or focus-change event
will follow.

This commit eliminates that chattiness by waiting for the subsequent event.