Web: Handle text insertion events preceding focus events for entries
When navigating amongst entries, we ignore the text-selection-changed
events if we haven't yet received the focused event for that entry.
We need to do something similar in the case of entries whose text is
populated immediately before focus for two reasons: 1) The default
script silently updates the locus of focus which means the role is
not announced, but the text is still spoken. 2) If we then get a
text-selection-changed event, we present it again because it now is
coming from the locus of focus. Waiting until we get the focus event
ensures we announce the role, the text contents, and any selection
present.