Web: Don't attempt to infer labels for ARIA widgets
Authors using ARIA are expected to provide names for their widgets,
and typically do. However, when they don't, and we try to infer what
might be the label based on surrounding text, we sometimes wind up
duplicating a string the author plans to present some other way.
And label inference is not especially performant. So only do the work
for native HTML widgets which lack an accessible name.