Add ability for scripts to check if there's a pending event
Scripts currently have the ability to skip events (e.g. due to a
more-recent relevant event having been queued). This commit adds
the ability for scripts to check if their associated app/toolkit
has since fired a specified event. The main use case is related
to what to present and/or how to recover if we're processing an
event or generating speech/braille on an object that was destroyed
in the meantime. Using this new functionality will be done in
subsequent commits.