Web: Include focusable elements with clickancestor action as clickable
In order to minimize spam in the list of clickables, we were only
treating an element as clickable if it implemented the "click" action
and was not focusable. This made all the divs, spans, etc. that only
had a click handler discoverable and more accessible without producing
large lists of clickable items and saying "clickable" all the time. But
it causes items which have a clickable ancestor (exposed by Chromium)
to be ignored. A lot of things can have a clickable ancestor, so
only include them if they are also focusable.