Remove the not-focusable requirement for clickables in collection match rule
This check is redundant to the one in the predicate. So there should be
no noticeable change in the user experience from this commit.

In addition, there are some focusable elements which we want to treat as
clickable objects. Making those changes will be done in subsequent commit(s)
so that any regressions in the user experience can be reverted on a case-
by-case basis.