gnome-shell: fix regression related to not saying "window"
In order to be less chatty, Orca no longer speaks nameless windows
if it is about to speak something in that window as a result of a
queued up focused event. It is possible, however, for that event
to be for the nameless window, which we don't present because it
is already the focused object. Therefore, check that the queued up
focused event is not for the window which just became activated.

See issue #226