Web: Improve recovery related to locusOfFocus destruction
We attempt to recover when the currently-focused object is destroyed.
That logic relies on some equality checks, namely that the event.any_data
contains the same object as our current focus. Unfortunately, at least in
the case of Firefox, those checks can fail. It appears that in some cases
a new accessible object is being created for the purpose of firing the
event for that object's removal.

This commit solves that problem by calling isSameObject and doing a
path comparison. In addition, when doing the path comparison, it
strips off the final -1 child index when it is present in both paths.
The theory being that if the paths are otherwise valid and identical,
and the roles and names match, the objects are probably the same.