Web: Improve behavior when "activating" an element in the nav list
Orca's navlist was assuming that if an action was present, invoking
it would actually cause that action to be performed. This does not
appear to be the case with Chromium's "clickAncestor" action. Therefore,
if we see that action and/or the action interface exposes 0 actions,
attempt to synthesize an actual mouse click on the object.