Handle WebKitGtk's change in toolkit name casing
WebKitGtk objects now expose the toolkit name as WebKitGTK. Our
script mapping for toolkits should probably not be case sensitive.
For the purpose of making newer releases of WebKitGtk still work
with older, stable versions of Orca, this commit just handles
the alternative casing for WebKitGtk.

Note that in some cases (e.g. Yelp), native caret navigation via
F7 does not seem to be reliably toggled. That is a feature of either
the app or the toolkit; not of Orca. This commit does not impact that.

See issue #244