Don't depend on mouse release being emitted prior to caret-moved event
When the user clicks and releases the primary mouse button and a caret-
moved event is subsequently received, we present the line with the
caret. However, we sometimes see the caret-moved event prior to the
mouse-release event being fired. Therefore, if the caret moves and
the last input event was either primary mouse click or primary mouse
release, present the line.